Correcting inspection ability (#15)
* Modifying left and right actions input

* Changing inputs for inspected item's observation

* Fixing roll rotation

* Adding a transform variable for inspectable objects

This variable defines the transform of an object whenever the player inspects it.

* Adding smooth animation when starting inspection

The inspected item moves from its original location to us.

* Fixing animation and removing useless functions

Instead of copying the item's mesh, we move it towards the player. When the player stops inspecting, the initial transform is reapplied to the item.

* Hiding initial transform of the item

* Minor change on the animation

* Changing zoom

Zooming doesn't change the scale of the item anymore. Now, it only adjusts its distance with the player.

* Fixing zoom limits

---------
Co-authored-by: Adrien Philippon <57765484+adrimyla@users.noreply.github.com>