Adding a boolean preventing from picking up the controller too early
The player was able to pick up the controller without solving the puzzle in the hall. All they needed to do is to climb on the rocks below the pillar, then jump until the target detection raycast hit the controller.
I added a boolean variable to BP_ControllerPower, in order to prevent any interaction and highlight with the controller before solving the puzzle. After it's solved, I toggle this variable to let the player pick up the controller.